refactor: structure 파일들 합치기
