--- MusicDiary/MusicDiary/ContentData.swift
@@ -1,18 +1,0 @@
-//
-//  ContentData.swift
-//  MusicDiary
-//
-//  Created by 1v1 on 2021/02/01.
-//
-
-import UIKit
-
-struct ContentData {
-    var authorID: String?
-    var conentText: String?
-    var musicTitle: String?
-    var musicArtist: String?
-    var musicCoverUrl: URL?
-    var date: Date?
-    
-}
