--- MusicDiary/MusicDiary/DiaryStructure.swift
@@ -1,20 +1,0 @@
-//
-//  DiaryStructure.swift
-//  MusicDiary
-//
-//  Created by 1v1 on 2021/02/01.
-//
-
-import UIKit
-
-struct DiaryStructure {
-    var diaryId: String?
-    var contentList = [ContentData]()
-    var diaryName: String?
-    var diaryImageUrl: URL?
-    var diaryMusicTitle:String?
-    var diaryMusicArtist: String?
-    var memberList: [String]?
-    var date: Date?
-    
-}
