--- MusicDiary/MusicDiary/UserStructure.swift
@@ -1,15 +1,0 @@
-//
-//  UserStructure.swift
-//  MusicDiary
-//
-//  Created by 1v1 on 2021/01/31.
-//
-
-import UIKit
-
-public struct UserStructure {
-    var userName: String?
-    var userId: String?
-    var userImage: URL?
-    var userDiaryList: [String]?
-}
